\docType{data}
\name{amlsample}
\alias{amlsample}
\alias{aml.sample}
\title{Multitube AML sample as example data for flowBin}
\format{a \code{flowSample} containing 7 tubes with 3 common parameters and 4 measure parameters per tube.}
\source{
  FlowRepository.org accession FR-FCM-ZZYA
}
\description{
  Multitube AML sample as example data for flowBin
}
\keyword{datasets}

