\name{BinnedFlowSample}
\alias{bin.labels}
\alias{bin.labels<-}
\alias{bin.labels<-,BinnedFlowSample,list-method}
\alias{bin.labels,BinnedFlowSample-method}
\alias{BinnedFlowSample}
\alias{BinnedFlowSample-class}
\title{A \code{FlowSample}, but with binning information for each tube}
\description{
  \describe{ \item{\code{clust.labels}:}{\code{list} of
  cluster label vectors, one for each tube} Note: all slots
  can be get and set using accessor methods, for example
  \code{bin.pars(myFlowSet) <- c(1,2,5)} }
}

