\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show,BinnedFlowExprSet-method}
\alias{show,CVResult-method}
\title{ Methods to view flowBin objects }
\description{
Methods for function \code{show} in Package flowBin 
}
\section{Methods}{
\describe{

\item{\code{signature(object = "BinnedFlowExprSet")}}{
Show number of bins and samples for a BinnedFlowExprSet.
}

\item{\code{signature(object = "CVResult")}}{
Show various statistics stored in a CVResult object.
}
}}
\keyword{methods}
